MTC-10174 refactor mjml theme tokens
