+++ plugins/GrapesJsBuilderBundle/Assets/library/js/plugins/grapesjs-mjmlThemeTokens/utils.js
@@ -1,0 +1,26 @@
+import mjml2html from 'mjml-browser';
+
+export const pluginId = 'grapesjs-mjml-theme-tokens';
+
+export const extractMjHeadContent = (mjml) => {
+  if (!mjml) return '';
+  const m = mjml.match(/<mj-head[^>]*>([\s\S]*?)<\/mj-head>/i);
+  return m && m[1] ? m[1].trim() : '';
+};
+
+export const createHeadInjectingMjmlParser = (headContent = '') => (input, opts) => {
+  if (typeof input !== 'string') {
+    return mjml2html(input, opts);
+  }
+
+  if (!headContent || !/<mjml[\s>]/i.test(input) || /<mj-head[\s>]/i.test(input)) {
+    return mjml2html(input, opts);
+  }
+
+  const withHead = input.replace(
+    /<mjml(\s[^>]*)?>/i,
+    (m) => `${m}<mj-head>${headContent}</mj-head>`
+  );
+
+  return mjml2html(withHead, opts);
+};
